requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
redis==3.5.2
requests==2.26.0
rq==1.8.1
urllib3==1.26.7
Werkzeug==2.0.2
SQLAlchemy~=1.4.27
python-dotenv==0.19.2
redbaron==0.9.2
amazon-braket-sdk==1.35.5
gunicorn
pytz~=2023.3
alembic~=1.10.2